pyproject.toml
[project]
name = "rabbitmq-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
